[PATCH] grilo: don't hammer tracker with a query for every album
If tracker didn't give us a thumbnail the first time, it's because
it doesn't have one, so there is no point in querying again (especially
if we're not really querying for the thumbnail, but only for
artist and album fields, which we already know!). Instead, what
we want is to resolve the item using the other sources (GRL_RESOLVE_FULL),
in particular grl-local-metadata and grl-lastfm-albumart.

https://bugzilla.gnome.org/show_bug.cgi?id=706027